Refactor code, add path

+++ src/components/users/DeleteUserModal.vue
@@ -1,54 +1,58 @@
 <script setup lang="ts">
 import { computed } from "vue";
+import { storeToRefs } from "pinia";
 import { VueFinalModal } from "vue-final-modal";
 import { useUsersStore } from "../../store/users";
 import type { User } from "../../types/User";
 import { useFetch } from "../../composables/useFetch";
 import Btn from "@/components/Btn.vue";
 
-defineProps<{
-  title?: string;
-}>();
-
 const usersStore = useUsersStore();
+const { getDeleteUser } = storeToRefs(usersStore);
 
-const handleConfirm = () => {
-  const { id } = usersStore.getDeleteUser;
-  useFetch(
+const handleCloseModal = () => {
+  usersStore.deleteUserModalState = false;
+  usersStore.deleteUser = {} as User;
+};
+
+const handleConfirm = async () => {
+  const { id } = getDeleteUser.value;
+
+  if (!id) {
+    return;
+  }
+
+  await useFetch(
     `users/${id}`,
     {},
     { method: "DELETE" },
     "The user has been deleted."
   );
-  usersStore.deleteUserModalState = false;
-  usersStore.deleteUser = {} as User;
-};
 
-const handleCancel = () => {
-  usersStore.deleteUserModalState = false;
-  usersStore.deleteUser = {} as User;
+  handleCloseModal();
 };
 
 const userData = computed(() => {
-  const { first_name, last_name } = usersStore.getDeleteUser;
+  const { first_name, last_name } = getDeleteUser.value;
   return `${first_name} ${last_name}`;
 });
 </script>
 
 <template>
   <VueFinalModal
+    v-model="usersStore.deleteUserModalState"
     class="confirm-modal"
     content-class="confirm-modal-content"
     overlay-transition="vfm-fade"
     content-transition="vfm-fade"
-    @closed="handleCancel"
+    @closed="handleCloseModal"
   >
     <h2>Delete user</h2>
     <p>Are you sure you want to delete: {{ userData }}?</p>
 
     <div class="flex flex-wrap gap-4">
       <Btn @click="handleConfirm">YES</Btn>
-      <Btn @click="handleCancel" btnStyle="outline">NO</Btn>
+      <Btn @click="handleCloseModal" btnStyle="outline">NO</Btn>
     </div>
   </VueFinalModal>
 </template>
